Remove v1 splus backend
