+++ server/App.ts
@@ -1,9 +1,8 @@
 import * as express from 'express';
-import splusController from './controllers/v1/splus';
+import splusController from './controllers/splus';
 import mensaController from './controllers/mensa';
 import newsController from './controllers/news';
 import icsController from './controllers/ics';
-import splusController2 from './controllers/v2/splus';
 
 class App {
   public app: express.Application;
@@ -18,7 +17,6 @@ class App {
     this.app.use(this.basePath + '/mensa', mensaController);
     this.app.use(this.basePath + '/news', newsController);
     this.app.use(this.basePath + '/ics', icsController);
-    this.app.use(this.basePath + '/v2/splus', splusController2);
   }
 }
