+++ server/controllers/ics.ts
@@ -4,8 +4,8 @@ import * as ical from 'ical-generator';
 import { createHash } from 'crypto';
 
 import * as TIMETABLES from '../../assets/timetables.json';
-import { Event, TimetableRequest } from '../model/v2/SplusEinsModel';
-import getEvents from '../lib/v2/SplusApi';
+import { Event, TimetableRequest } from '../model/SplusEinsModel';
+import getEvents from '../lib/SplusApi';
 
 const router = express.Router();
 const sha256 = (x) => createHash('sha256').update(x, 'utf8').digest('hex');
