--- server/lib/v1/ILecture.ts
@@ -1,13 +1,0 @@
-export interface ILecture {
-    title: string;
-    day: number;
-    begin: number;
-    end: number;
-    info: string;
-    room: string,
-    lecturer: string;
-}
-
-export interface ILectureFilter {
-    (lecture: ILecture): boolean;
-}
