+++ server/model/SplusEinsModel.ts
