+++ server/model/SpluseEinsEvent.spec.ts
